bro era sparita tutta la roba che utilizza MonthlyReservations

+++ src/main/java/it/unipi/erasmusnest/controllers/MyReservationsController.java
@@ -16,6 +16,7 @@ import javafx.scene.layout.VBox;
 import java.nio.file.Path;
 import java.time.LocalDate;
 import java.util.ArrayList;
+import java.util.List;
 import java.util.Objects;
 
 public class MyReservationsController extends Controller {
@@ -34,17 +35,43 @@ public class MyReservationsController extends Controller {
 
         System.out.println("MyReservations initialize");
 
-        reservations = getRedisConnectionManager().getReservationsForUser(getSession().getUser().getEmail());
-        if(!reservations.isEmpty()){
-            for (Reservation reservation : reservations) {
-                System.out.println(reservation.toString());
-                add(reservation);
+        if(getSession().getApartmentsId() == null){
+            reservations = getRedisConnectionManager().getReservationsForUser(getSession().getUser().getEmail());
+            if(!reservations.isEmpty()){
+                for (Reservation reservation : reservations) {
+                    add(reservation, "student");
+                }
+            } else {
+                noReservation();
             }
         } else {
-            noReservation();
+            // get the reservations for the apartments
+            List<Long> apartmentsId = getSession().getApartmentsId();
+            reservations = getRedisConnectionManager().getReservationsForApartments(apartmentsId);
+
+            // ordering the reservations by timestamp
+            sortReservationsByTimestampAsc();
+
+            for(Reservation reservation : reservations){
+                System.out.println(">>> "+reservation.getApartmentId()+" "+reservation.getTimestamp());
+                add(reservation, "host");
+            }
+
         }
 
         scrollPane.setVvalue(0.0);
+    }
+
+    private void sortReservationsByTimestampAsc(){
+        reservations.sort((o1, o2) -> {
+            if(o1.getTimestamp().isBefore(o2.getTimestamp())){
+                return -1;
+            } else if(o1.getTimestamp().isAfter(o2.getTimestamp())){
+                return 1;
+            } else {
+                return 0;
+            }
+        });
     }
 
     private void noReservation(){
@@ -62,7 +89,9 @@ public class MyReservationsController extends Controller {
 
     }
 
-    private void add(Reservation reservation) {
+    private void add(Reservation reservation, String userType) {
+
+        // userType = "host" | "student"
 
         HBox reservationHBox = new HBox();
         reservationHBox.setAlignment(Pos.CENTER_LEFT);
@@ -75,6 +104,8 @@ public class MyReservationsController extends Controller {
         cityVBox.setAlignment(javafx.geometry.Pos.CENTER);
         VBox periodVBox = new VBox();
         periodVBox.setAlignment(javafx.geometry.Pos.CENTER);
+        VBox stateVBox = new VBox();
+        stateVBox.setAlignment(javafx.geometry.Pos.CENTER);
         VBox buttonsVBox = new VBox();
         buttonsVBox.setAlignment(javafx.geometry.Pos.CENTER);
 
@@ -117,32 +148,73 @@ public class MyReservationsController extends Controller {
         Label period = new Label(periodFromTo);
         period.setStyle("-fx-font-size: 15px; -fx-text-fill: #ff6f00;");
 
-        imageVBox.getChildren().add(imageView);
-        cityVBox.getChildren().add(city);
-        periodVBox.getChildren().add(period);
+        String state = "state: "+reservation.getState();
+        Label stateLabel = new Label(state);
+        stateLabel.setStyle("-fx-font-size: 15px; -fx-text-fill: #019fe1;");
 
         // remove '\n' from periodFromTo
         String msgPeriod = periodFromTo.replace("\n", " ");
-        Button deleteButton = new Button("Delete reservation");
-        deleteButton.setStyle("-fx-background-color: #ff0000; -fx-text-fill: #ffffff; -fx-font-size: 11px; -fx-font-weight: bold; -fx-background-radius: 5px;");
-        deleteButton.setOnAction(event -> {
-            boolean remove = new AlertDialogGraphicManager("Are you sure you want to delete this reservation\n"
-                    + msgPeriod +" in "+ reservation.getCity()
-                    +"?", "You will not be able to recover it").showAndGetConfirmation();
-            if(remove){
-                getRedisConnectionManager().deleteReservation(reservation);
-                super.changeWindow("myreservations");
-            }
-        });
-        buttonsVBox.getChildren().add(deleteButton);
+
+        buildButtons(buttonsVBox, reservation, userType, msgPeriod);
+
+        imageVBox.getChildren().add(imageView);
+        cityVBox.getChildren().add(city);
+        periodVBox.getChildren().add(period);
+        stateVBox.getChildren().add(stateLabel);
 
         imageVBox.prefWidthProperty().bind(reservationHBox.widthProperty().multiply(0.5));
         cityVBox.prefWidthProperty().bind(reservationHBox.widthProperty().multiply(0.5));
         periodVBox.prefWidthProperty().bind(reservationHBox.widthProperty().multiply(0.5));
+        stateVBox.prefWidthProperty().bind(reservationHBox.widthProperty().multiply(0.5));
         buttonsVBox.prefWidthProperty().bind(reservationHBox.widthProperty().multiply(0.5));
-        reservationHBox.getChildren().addAll(imageVBox, cityVBox, periodVBox, buttonsVBox);
+        reservationHBox.getChildren().addAll(imageVBox, cityVBox, periodVBox, stateVBox, buttonsVBox);
         centerVBox.getChildren().add(reservationHBox);
 
+    }
+
+    private void buildButtons(VBox buttonsVBox, Reservation reservation, String userType, String msgPeriod){
+        if(userType.equals("student")) {
+            Button deleteButton = new Button("Delete reservation");
+            deleteButton.setStyle("-fx-background-color: #ff0000; -fx-text-fill: #ffffff; -fx-font-size: 11px; -fx-font-weight: bold; -fx-background-radius: 5px;");
+            deleteButton.setOnAction(event -> {
+                boolean remove = new AlertDialogGraphicManager("Are you sure you want to delete this reservation\n"
+                        + msgPeriod + " in " + reservation.getCity()
+                        + "?", "You will not be able to recover it").showAndGetConfirmation();
+                if (remove) {
+                    getRedisConnectionManager().deleteReservation(reservation);
+                    super.changeWindow("myreservations");
+                }
+            });
+            buttonsVBox.getChildren().add(deleteButton);
+        } else if(reservation.getState().equals("pending")){
+            Button approveButton = new Button("Approve reservation");
+            approveButton.setStyle("-fx-background-color: #63d27f; -fx-text-fill: #ffffff; -fx-font-size: 11px; -fx-font-weight: bold; -fx-background-radius: 5px;");
+            approveButton.setOnAction(event -> {
+                boolean approve = new AlertDialogGraphicManager("Are you sure you want to approve this reservation\n"
+                        + msgPeriod + " in " + reservation.getCity()
+                        + "?", "You will not be able to reject it later").showAndGetConfirmation();
+                if (approve) {
+                    // TODO getRedisConnectionManager().approveReservation(reservation);
+                    super.changeWindow("myreservations");
+                }
+            });
+            Button rejectButton = new Button("Reject reservation");
+            rejectButton.setStyle("-fx-background-color: #ff0000; -fx-text-fill: #ffffff; -fx-font-size: 11px; -fx-font-weight: bold; -fx-background-radius: 5px;");
+            rejectButton.setOnAction(event -> {
+                boolean approve = new AlertDialogGraphicManager("Are you sure you want to reject this reservation\n"
+                        + msgPeriod + " in " + reservation.getCity()
+                        + "?", "You will not be able to approve it later").showAndGetConfirmation();
+                if (approve) {
+                    // TODO getRedisConnectionManager().rejectReservation(reservation);
+                    super.changeWindow("myreservations");
+                }
+            });
+            VBox vBox = new VBox();
+            vBox.setAlignment(Pos.CENTER);
+            vBox.setSpacing(10);
+            vBox.getChildren().addAll(approveButton, rejectButton);
+            buttonsVBox.getChildren().add(vBox);
+        }
     }
 
     @FXML
